auth completed little refactoring
